+++ app/modules/User/controller.js
@@ -4,22 +4,44 @@
 'use strict'
 
 const User = require('./model')
+const faker = require('faker')
 
 function index (request, response) {
   /**
    * This is only an example.
-   * You need to create an user in your database.
+   * You can access `/users/create` to create an random user
    */
-  User.findOne({}, (err, user) => {
-    if (err) {
-      return response.status(500).send(err)
-    }
+  User.find()
+    .then(users => {
+      if (!users) {
+        return response.send('no users')
+      }
 
-    response.format({
-      html: () => response.render('User/index', {user: user}),
-      json: () => response.json(user)
+      response.format({
+        html: () => response.render('User/index', {users: users}),
+        json: () => response.json(users)
+      })
     })
-  })
+    .catch(err => response.status(500).json(err))
 }
 
-module.exports = {index}
+/**
+ * Create.
+ *
+ * This method create an random user.
+ */
+function create (request, response) {
+  let user = new User({
+    name: faker.name.firstName(),
+    email: faker.internet.email(),
+    username: faker.internet.userName(),
+    password: faker.lorem.word()
+  })
+
+  user
+    .save()
+    .then(user => response.redirect('/users'))
+    .catch(err => response.status(500).json(err))
+}
+
+module.exports = {index, create}