+++ app/modules/User/routes.js
@@ -7,5 +7,6 @@ const user = require('./controller')
 const router = require('express').Router()
 
 router.get('/', user.index)
+router.get('/create', user.create)
 
 module.exports = router