+++ app/routes.js
@@ -1,12 +1,12 @@
 /**
- * Router
+ * Routes
  */
 'use strict'
 
 const user = require('./modules/User')
 
 module.exports = app => {
-  app.use('/user', user)
+  app.use('/users', user)
 
   app.get('/', (request, response) => {
     response.render('../views/welcome')