+++ config/database.js
@@ -3,8 +3,10 @@
  */
 'use strict'
 
+require('dotenv').config()
+
 module.exports = {
   mongo: {
-    uri: 'mongodb://localhost:28002/homestead'
+    uri: process.env.MONGODB_URI
   }
 }