+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Unreleased
 
+- Create helpers function
+- Add crypt for passwords
+- Auth example
+- Configure connection to connect only when this is used
+
+## [0.1.0] - YYYY-MM-DD
+
+### Added
+- Add .env files
 - Added public to static files
 - Added main route file
 - Added route file to module